# perf: cache css value