before

React#renderToString x 1,170 ops/sec ±1.24% (86 runs sampled)
Rax#renderToString x 2,032 ops/sec ±1.24% (89 runs sampled)
Inferno#renderToString x 4,121 ops/sec ±1.06% (88 runs sampled)

after

React#renderToString x 1,177 ops/sec ±1.28% (83 runs sampled)
Rax#renderToString x 4,908 ops/sec ±1.49% (83 runs sampled)
Inferno#renderToString x 3,649 ops/sec ±10.91% (79 runs sampled)